Robot-KG/requirements.txt
torch==1.11
hydra-core==1.0.6
tensorboard==2.4.1
matplotlib==3.4.1
transformers==3.4.0
jieba==0.42.1
scikit-learn==0.24.1
seqeval==1.2.2
tqdm==4.60.0
opt-einsum==3.3.0
wandb==0.12.7
ujson==5.4.0
huggingface_hub==0.2.1
elasticsearch==6.8.0
flask==2.2.2
py2neo==2021.2.3
cachetools==4.2.4
deepke==2.1.0
google-auth==1.35.0
ipdb==0.13.9
numpy==1.23.2
pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.20.0
ptyprocess==0.7.0
pytorch-crf==0.7.2
requests==2.28.1
setuptools==62.4
tokenizers==0.9.2
wheel==0.37.1
tk
tensorboardX==2.5.1
pandas==1.5.0
openpyxl==3.0.10
setuptools-rust==1.5.2
pyyaml
flask_cors==3.0.10